Power packet optimization
By increasing the number of display stages to 100, I was able to easily
map internal power to display stage by sending a byte representing the
display stage rather that the amount of power in the pipe. This allows
me to reduce the packet size in half by converting from Shorts to Bytes.

Additionally, I created an alternative display list that uses a red
texture to indicate that pipe is nearing maximum capacity.
Unfortunately, this still gives no visual cue that a wooden pipe is
about to explode, becasue for some reason wooden pipes don't render the
power they contain when connected to an engine. This should probably be
rectified in the future.