Added Emerald  Item/Liquid Pipes
Also new texture for the Laser Crafting Table and fixed a bug with
Wooden Item Pipes not diplaying items that are being pumped.

Emerald Pipes are advanced versions of the Wooden Pipes. The Item Pipe
posseses a filter that pulls items from inventories in a blocking Round
Robin fashion. The Liquid Pipe has capacity and flow rate equal to a
golden pipe.

Note: ISpecial and ISelective Inventory interaction with the Emerald
Item Pipes is untested, there seems to be nothing decent to test with in
Buildcraft.