Wooden Power Pipes now add power to themselves...
...instead of their neighbor. The result being that they are constrained
by the new limits and that they actually render the power flow in
themselves.

Wooden Pipes are set with a input cap of 32 MJ/t, buffed from the
previous effective limit of 29 MJ/t.

Note that this also effectively adds a 32 MJ/t pipe that can be used for
limiting flows.