Fixed Power Request Trigger
Its now restricted to Wooden Power pipes.

Before:
The trigger was useless because it was triggering off the powerQuery
array which was always true so long as there was any device on the
network requesting power. This made it rather useless for scaling
production to demand.

After:
It now triggers whenever the Wooden Power Pipe's internal power buffer
drops below a certain point. This however required a modification of the
wooden pipe's internals because it was getting stuck in the "Off" state
due to how it handled transferring power from the buffer to the pipe
network.

Anyway, it works now and is somewhat useful.

Note: The Wooden Power Pipe's internal buffer stores 1500 MJ, this was
true even before this commit.