Improve Pump search Algo
Mainly just cleaned up the code and replaced the TreeSets with Deques.

Also made it so the pump will only keep the pumping the same type of
liquid it started on.  If you wish to reset it to pump any liquid hit
with a wrench and the next liquid it pumps will be the new filter.