Fix Pipe Raytracing + Improve Pipe interactions
Closes #1246

Its now much easier and more consistent to remove specific addons from a
pipe. Sneak-click with a gate to remove a gate, sneak-click with a
Yellow Wire to remove a Yellow Wire, etc...

Plugs can now be placed easily and simply on any side of the pipe in any
configuration.

Selection Box now highlights the part of the pipe you are pointing at.